Start attributing mixin locations to the last require location
We have been trying to figure out the location of a mixin by
introspecting the full backtrace of the point of mixin. We were
assuming that if we found the gem path in the backtrace at any position
that meant that the gem was responsible for the mixin.

However, that assumption breaks when the type that triggers the mixin
is ultimately different from the file that actually does the mixin.

For example, in Rails, controllers can call `helper :Foo` to make
`ActionController` create a `FooHelper` module and `include` it into
the original controller. In this case, the location closest to the
actual `include` location is inside the `actionpack` gem. However, we
don't want the attribution for the mixin to be for `actionpack`, we
want it to be for the controller that triggered the mixin by calling
the `helper` method.

The way we can figure out which gem we should attribute a mixin to, is
by looking at the file that was required that is closest to the mixin
location. This means that we walk the backtrace back until we find a
frame that refers to `"<top (required)>"` (a file being required), and
we can attribute the mixin to that location.

This commit extracts the logic to do that attribution into a method on
`Reflection` and starts using it in the `Mixin` tracker.