Refactor `ConstantDefinition` to use `required_from_location`
Now that we have a better way of attributing a backtrace to the gem
that triggered it, we should be using it in `ConstantDefinition` as
well.

Moreover, instead of storing constant names as keys in
`ConstantDefinition`, we should start storing the actual constant in a
compare by identity hash, so that both storage and lookups can be
faster.