Start tracking dynamic module definitions in `ConstantDefinition`
So far, we had only been tracking explicit class/module definitions,
i.e. classes/modules defined by using `class Foo` or `module Bar`.

However, we had no way of tracking constants that were created
dynamically using `Class.new` or `Module.new`. But, it is important to
track these constants to do correct constant attribution, since a lot
of constants inside Rails, specifically are generated dynamically.

For example:
```ruby
module Foo
  extend ActiveSupport::Concern

  class_methods {}
end
```
would create the module `Foo::ClassMethods` dynamically, inside
`activesupport`. We need a way to catch that definition and attribute it
to the gem that defines the `Foo` module.

The way to do that is by starting to listen to the `c_return` TracePoint
so that we can catch C-calls to `.new` methods which return `Module`
instances. By relying on `required_for_location` to do the correct
attribution, and building on the fact that we are using constants as
lookup keys now (since dynamically created modules might not have a
name), we can now store these kinds of constants in the same lookup
table that we have built.