Handle mixins into singleton classes
We need to handle mixins into singleton class, since:
```ruby
Foo.singleton_class.prepend(Bar)
```
is a common idiom. However, in this case, we end up with the singleton
class as the mixee and when we ask it for its name, we get `nil`. Thus,
we end up never considering the attached class for generation either.

So, if the name of the mixee is `nil`, we now try to parse the name of
the base class from the `to_s` representation of the name of the
singleton class and we queue that symbol up for generation. Since we
handle the mixins of both the base class and the singleton class
separately, this is guaranteed to generate the appropriate mixins on the
singleton class.